+++ modules/sdk/gradle-plugins-node/src/main/java/com/liferay/gradle/plugins/node/tasks/ExecuteNodeTask.java
@@ -15,8 +15,8 @@
 package com.liferay.gradle.plugins.node.tasks;
 
 import com.liferay.gradle.plugins.node.NodePlugin;
-import com.liferay.gradle.plugins.node.util.GradleUtil;
-import com.liferay.gradle.plugins.node.util.NodeExecutor;
+import com.liferay.gradle.plugins.node.internal.NodeExecutor;
+import com.liferay.gradle.plugins.node.internal.util.GradleUtil;
 
 import java.io.File;
 import java.io.IOException;
