+++ modules/sdk/gradle-plugins-node/src/main/java/com/liferay/gradle/plugins/node/tasks/NpmInstallTask.java
@@ -14,8 +14,8 @@
 
 package com.liferay.gradle.plugins.node.tasks;
 
-import com.liferay.gradle.plugins.node.util.FileUtil;
-import com.liferay.gradle.plugins.node.util.GradleUtil;
+import com.liferay.gradle.plugins.node.internal.util.FileUtil;
+import com.liferay.gradle.plugins.node.internal.util.GradleUtil;
 
 import groovy.json.JsonSlurper;
